requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asgiref==3.8.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
clarifai==11.6.5
clarifai-grpc==11.6.5
clarifai-protocol==0.0.25
click==8.2.1
contextlib2==21.6.0
Django==5.2
django-allauth==65.8.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
ecdsa==0.19.1
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.1
googleapis-common-protos==1.70.0
grpcio==1.73.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
multidict==6.6.3
numpy==2.3.1
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-jose==3.5.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
schema==0.7.5
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tqdm==4.67.1
uritemplate==4.1.1
urllib3==2.4.0
uv==0.7.12
yarl==1.20.1
easyocr==1.7.1
